requirements.txt
asgiref==3.8.1
boto3==1.34.145
botocore==1.34.145
dj-database-url==0.5.0
Django==4.2.7
django-allauth==0.63.3
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.14.4
django-summernote==0.8.20.0
django-taggit==6.0.0
gunicorn==22.0.0
jmespath==1.0.1
pillow==10.3.0
psycopg2==2.9.9
pytz==2024.1
s3transfer==0.10.2
sqlparse==0.5.0
stripe==10.1.0
